+++ stories/index.js
@@ -47,4 +47,4 @@ require('./input');
 require('./icon');
 require('./select');
 require('./checkbox');
-require('./radioGroup')
+require('./radioButton')
